backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiodns==3.0.0
    # via aiohttp
aiohttp[speedups]==3.8.4
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.19.0
    # via -r requirements.in
alembic==1.10.3
    # via -r requirements.in
amqp==5.1.1
    # via kombu
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
asgiref==3.7.2
    # via -r requirements.in
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
asyncpg==0.27.0
    # via -r requirements.in
attrs==23.1.0
    # via aiohttp
billiard==4.1.0
    # via celery
black==23.3.0
    # via -r requirements.in
brotli==1.0.9
    # via aiohttp
build==0.10.0
    # via pip-tools
celery==5.3.4
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   pytest-celery
celery-redbeat==2.1.1
    # via -r requirements.in
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
cffi==1.15.1
    # via pycares
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.2.0
    # via aiohttp
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coverage[toml]==7.2.3
    # via pytest-cov
distlib==0.3.6
    # via virtualenv
dnspython==2.3.0
    # via email-validator
email-validator==2.0.0.post2
    # via pydantic
exceptiongroup==1.1.1
    # via pytest
execnet==1.9.0
    # via pytest-xdist
faker==18.7.0
    # via polyfactory
fakeredis==2.19.0
    # via -r requirements.in
fastapi==0.95.0
    # via
    #   -r requirements.in
    #   strawberry-graphql
filelock==3.12.0
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements.in
    #   pep8-naming
freezegun==1.2.2
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
graphql-core==3.2.3
    # via strawberry-graphql
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.2
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.1
    # via -r requirements.in
identify==2.5.24
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   yarl
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements.in
jmespath==1.0.1
    # via -r requirements.in
kombu==5.3.2
    # via celery
loguru==0.7.2
    # via -r requirements.in
lxml==4.9.3
    # via -r requirements.in
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via mako
mccabe==0.7.0
    # via flake8
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy==1.3.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
packaging==23.0
    # via
    #   black
    #   build
    #   pytest
    #   pytest-sugar
pathspec==0.11.1
    # via black
pep8-naming==0.13.3
    # via -r requirements.in
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.2.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
polyfactory==2.1.0
    # via -r requirements.in
pre-commit==3.3.2
    # via -r requirements.in
prompt-toolkit==3.0.39
    # via click-repl
pycares==4.3.0
    # via aiodns
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pydantic[email]==1.10.7
    # via
    #   -r requirements.in
    #   fastapi
    #   sqlmodel
pyflakes==3.0.1
    # via flake8
pyproject-hooks==1.0.0
    # via build
pytest==7.3.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-env
    #   pytest-mock
    #   pytest-sugar
    #   pytest-xdist
pytest-asyncio==0.21.0
    # via -r requirements.in
pytest-celery==0.0.0
    # via -r requirements.in
pytest-cov==4.0.0
    # via -r requirements.in
pytest-env==0.8.1
    # via -r requirements.in
pytest-mock==3.10.0
    # via -r requirements.in
pytest-sugar==0.9.7
    # via -r requirements.in
pytest-xdist==3.2.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   celery
    #   celery-redbeat
    #   faker
    #   freezegun
    #   strawberry-graphql
python-dotenv==1.0.0
    # via uvicorn
python-multipart==0.0.6
    # via strawberry-graphql
pyyaml==6.0
    # via
    #   pre-commit
    #   uvicorn
redis==5.0.0
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   fakeredis
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sortedcontainers==2.4.0
    # via fakeredis
sqlalchemy==1.4.41
    # via
    #   -r requirements.in
    #   alembic
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a34
    # via sqlmodel
sqlmodel==0.0.8
    # via -r requirements.in
starlette==0.26.1
    # via fastapi
strawberry-graphql[fastapi]==0.171.1
    # via -r requirements.in
tenacity==8.2.3
    # via celery-redbeat
termcolor==2.2.0
    # via pytest-sugar
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pyproject-hooks
    #   pytest
typing-extensions==4.5.0
    # via
    #   alembic
    #   asgiref
    #   mypy
    #   polyfactory
    #   pydantic
    #   sqlalchemy2-stubs
    #   strawberry-graphql
tzdata==2023.3
    # via celery
uvicorn[standard]==0.21.1
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.23.0
    # via pre-commit
watchfiles==0.19.0
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==11.0.1
    # via uvicorn
wheel==0.40.0
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
